+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Protocol is now at version **1.13**. See `API changelog`_.
 - Add missing ``ETag`` and ``Last-Modified`` headers on ``POST`` and ``DELETE``
   responses (#980)
 - Return 404 on non-existing objects for users with read permissions (fixes #918)
+- Fix pagination with DELETE on plural endpoints (fixes #987)
 
 **New features**